% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Bank}
\format{
A data frame with 100 rows and 2 variables (columns):
\describe{
\item{...1}{numerical vector counting the number of customers}
\item{arrival_time}{numerical vector containing arrival time (in seconds) of customers; 100 cumulated simulated variables from exponential distribution}
\item{service_time}{numerical vector containg service time (in seconds) of customers; 100 simulated variables from exponential distribution}
}
}
\usage{
bank
}
\description{
A data set which inclues the arrival and service times of customers at a bank
}
\keyword{datasets}
