% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Generate bank data}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{This is the number of arrivals at a bank.}

\item{ServiceTimes}{This is the service times (start and end time of service) of customers at a bank.}

\item{NumServers}{This is the number of servers at a bank.}
}
\value{
Returns a tibble which includes the arrivals, start and end of service times of the customers at a bank.
}
\description{
This generates data of customers' arrival and service times at a bank.
}
\examples{
Multiserver(Arrivals = 1, ServiceTimes = 2, NumServers = 1)
}
